+++ src/main/java/com/epam/reportportal/auth/integration/github/GitHubTokenServices.java
@@ -40,6 +40,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.function.Supplier;
 
+import static java.util.Collections.emptyList;
 import static java.util.Optional.ofNullable;
 
 /**
@@ -63,8 +64,9 @@ public class GitHubTokenServices implements ResourceServerTokenServices {
 		UserResource gitHubUser = gitHubClient.getUser();
 
 		List<String> allowedOrganizations = ofNullable(loginDetails.get().getRestrictions())
-				.flatMap(restrictions -> ofNullable(restrictions.get("organizations"))).map(it -> Splitter.on(",").splitToList(it))
-				.orElse(Collections.emptyList());
+				.flatMap(restrictions -> ofNullable(restrictions.get("organizations")))
+				.map(it -> Splitter.on(",").omitEmptyStrings().splitToList(it))
+				.orElse(emptyList());
 		if (!allowedOrganizations.isEmpty()) {
 			boolean assignedToOrganization = gitHubClient.getUserOrganizations(gitHubUser).stream().map(userOrg -> userOrg.login)
 					.anyMatch(allowedOrganizations::contains);
@@ -79,7 +81,7 @@ public class GitHubTokenServices implements ResourceServerTokenServices {
 				AuthUtils.AS_AUTHORITIES.apply(user.getRole()));
 
 		Map<String, Serializable> extensionProperties = Collections.singletonMap("upstream_token", accessToken);
-		OAuth2Request request = new OAuth2Request(null, null, null, true, null, null, null, null, extensionProperties);
+		OAuth2Request request = new OAuth2Request(null, loginDetails.get().getClientId(), null, true, null, null, null, null, extensionProperties);
 		return new OAuth2Authentication(request, token);
 	}