+++ CHANGELOG.md
@@ -1,15 +1,25 @@
 # Changelog
 
-## 2.6.0-SNAPSHOT
-##### Released: XXX September 2016
+## 2.6.0
+##### Released: October, 2016
 
 ### New Features
 
 * Initial release to Public Maven Repositories
 * DockerHub Release 
 
+
+## 2.7.0
+##### Released: XXX 2017
+
+### New Features
+
+* EPMRPP-20577,EPMRPP-20750 - GitHUB settings are moved to database and made configurable through UI
+* EPMRPP-23044 - GitHUB user synchronization
+* EPMRPP-23477 - Added GitHub integration status to UAT's service info endpoint
+
 ### Bugfixes
 
-* XXX
+* EPMRPP-23535 - If GitHub organization is empty, no one can log in